# apps/moment/public/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "My Moment API",
    "version": "1.0.0",
    "description": "Read and manage photos and tags. Every route requires the generated API key."
  },
  "servers": [
    {
      "url": "https://moment.you-find.me/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
    "/photos": {
      "get": {
        "summary": "List or search photos",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 25,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "name": "tagMode",
            "in": "query",
            "required": false,
            "schema": {
              "default": "any",
              "type": "string",
              "enum": ["any", "all"]
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": ["asc", "desc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Photos, newest first. Pass nextCursor as cursor for the next page.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "photos": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          },
                          "thumbnailUrl": {
                            "type": "string"
                          },
                          "r2Key": {
                            "type": "string"
                          },
                          "thumbnailR2Key": {
                            "type": "string"
                          },
                          "thumbHash": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "width": {
                            "type": "number"
                          },
                          "height": {
                            "type": "number"
                          },
                          "aspectRatio": {
                            "type": "number"
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "date": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "size": {
                            "type": "number"
                          },
                          "format": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          },
                          "geo": {
                            "type": "object",
                            "properties": {
                              "lat": {
                                "type": "number",
                                "minimum": -90,
                                "maximum": 90
                              },
                              "lng": {
                                "type": "number",
                                "minimum": -180,
                                "maximum": 180
                              }
                            },
                            "required": ["lat", "lng"],
                            "additionalProperties": false
                          }
                        },
                        "required": [
                          "id",
                          "url",
                          "thumbnailUrl",
                          "r2Key",
                          "thumbnailR2Key",
                          "title",
                          "width",
                          "height",
                          "tags",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": {
                      "type": ["string", "null"]
                    }
                  },
                  "required": ["photos", "nextCursor"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a photo from objects already uploaded to R2",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "r2Key": {
                    "type": "string"
                  },
                  "thumbnailR2Key": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "tags": {
                    "default": [],
                    "maxItems": 10,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 50
                    }
                  },
                  "date": {
                    "type": "string"
                  },
                  "geo": {
                    "type": "object",
                    "properties": {
                      "lat": {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                      },
                      "lng": {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                      }
                    },
                    "required": ["lat", "lng"]
                  },
                  "thumbHash": {
                    "type": "string"
                  },
                  "width": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "height": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "aspectRatio": {
                    "type": "number",
                    "exclusiveMinimum": 0
                  },
                  "format": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": ["r2Key", "thumbnailR2Key", "title", "width", "height"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created photo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "photo": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "thumbnailUrl": {
                          "type": "string"
                        },
                        "r2Key": {
                          "type": "string"
                        },
                        "thumbnailR2Key": {
                          "type": "string"
                        },
                        "thumbHash": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "width": {
                          "type": "number"
                        },
                        "height": {
                          "type": "number"
                        },
                        "aspectRatio": {
                          "type": "number"
                        },
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "date": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "size": {
                          "type": "number"
                        },
                        "format": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        },
                        "geo": {
                          "type": "object",
                          "properties": {
                            "lat": {
                              "type": "number",
                              "minimum": -90,
                              "maximum": 90
                            },
                            "lng": {
                              "type": "number",
                              "minimum": -180,
                              "maximum": 180
                            }
                          },
                          "required": ["lat", "lng"],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "id",
                        "url",
                        "thumbnailUrl",
                        "r2Key",
                        "thumbnailR2Key",
                        "title",
                        "width",
                        "height",
                        "tags",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": ["photo"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Photo or R2 object not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "An R2 object is already used by another photo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/photos/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "Get a photo",
        "responses": {
          "200": {
            "description": "Photo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "photo": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "thumbnailUrl": {
                          "type": "string"
                        },
                        "r2Key": {
                          "type": "string"
                        },
                        "thumbnailR2Key": {
                          "type": "string"
                        },
                        "thumbHash": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "width": {
                          "type": "number"
                        },
                        "height": {
                          "type": "number"
                        },
                        "aspectRatio": {
                          "type": "number"
                        },
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "date": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "size": {
                          "type": "number"
                        },
                        "format": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        },
                        "geo": {
                          "type": "object",
                          "properties": {
                            "lat": {
                              "type": "number",
                              "minimum": -90,
                              "maximum": 90
                            },
                            "lng": {
                              "type": "number",
                              "minimum": -180,
                              "maximum": 180
                            }
                          },
                          "required": ["lat", "lng"],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "id",
                        "url",
                        "thumbnailUrl",
                        "r2Key",
                        "thumbnailR2Key",
                        "title",
                        "width",
                        "height",
                        "tags",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": ["photo"]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Photo or R2 object not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update photo metadata",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "tags": {
                    "maxItems": 10,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 50
                    }
                  },
                  "date": {
                    "type": ["string", "null"]
                  },
                  "geo": {
                    "anyOf": [
                      {
                        "type": "object",
                        "properties": {
                          "lat": {
                            "type": "number",
                            "minimum": -90,
                            "maximum": 90
                          },
                          "lng": {
                            "type": "number",
                            "minimum": -180,
                            "maximum": 180
                          }
                        },
                        "required": ["lat", "lng"]
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated photo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "photo": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "thumbnailUrl": {
                          "type": "string"
                        },
                        "r2Key": {
                          "type": "string"
                        },
                        "thumbnailR2Key": {
                          "type": "string"
                        },
                        "thumbHash": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "width": {
                          "type": "number"
                        },
                        "height": {
                          "type": "number"
                        },
                        "aspectRatio": {
                          "type": "number"
                        },
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "date": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "size": {
                          "type": "number"
                        },
                        "format": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        },
                        "geo": {
                          "type": "object",
                          "properties": {
                            "lat": {
                              "type": "number",
                              "minimum": -90,
                              "maximum": 90
                            },
                            "lng": {
                              "type": "number",
                              "minimum": -180,
                              "maximum": 180
                            }
                          },
                          "required": ["lat", "lng"],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "id",
                        "url",
                        "thumbnailUrl",
                        "r2Key",
                        "thumbnailR2Key",
                        "title",
                        "width",
                        "height",
                        "tags",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": ["photo"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Photo or R2 object not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a photo and its stored objects",
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Photo or R2 object not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/tags": {
      "get": {
        "summary": "List tags with photo counts",
        "responses": {
          "200": {
            "description": "Tags, most used first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "count": {
                            "type": "integer"
                          }
                        },
                        "required": ["name", "count"],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": ["tags"]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}